refactor: reap related bits to root/reap.c
