+++ root.c
@@ -83,61 +83,6 @@ void free_file_node(w_root_t *root, struct watchman_file *file)
   free(file);
 }
 
-static bool root_has_subscriptions(w_root_t *root) {
-  bool has_subscribers = false;
-  w_ht_iter_t iter;
-
-  pthread_mutex_lock(&w_client_lock);
-  if (w_ht_first(clients, &iter)) do {
-    struct watchman_user_client *client = w_ht_val_ptr(iter.value);
-    w_ht_iter_t citer;
-
-    if (w_ht_first(client->subscriptions, &citer)) do {
-      struct watchman_client_subscription *sub = w_ht_val_ptr(citer.value);
-
-      if (sub->root == root) {
-        has_subscribers = true;
-        break;
-      }
-
-    } while (w_ht_next(client->subscriptions, &citer));
-  } while (!has_subscribers && w_ht_next(clients, &iter));
-  pthread_mutex_unlock(&w_client_lock);
-  return has_subscribers;
-}
-
-// This is a little tricky.  We have to be called with root->lock
-// held, but we must not call w_root_stop_watch with the lock held,
-// so we return true if the caller should do that
-bool consider_reap(struct write_locked_watchman_root *lock) {
-  time_t now;
-  w_root_t *root = lock->root;
-
-  if (root->idle_reap_age == 0) {
-    return false;
-  }
-
-  time(&now);
-
-  if (now > root->last_cmd_timestamp + root->idle_reap_age &&
-      (root->commands == NULL || w_ht_size(root->commands) == 0) &&
-      (now > root->last_reap_timestamp) &&
-      !root_has_subscriptions(root)) {
-    // We haven't had any activity in a while, and there are no registered
-    // triggers or subscriptions against this watch.
-    w_log(W_LOG_ERR, "root %.*s has had no activity in %d seconds and has "
-        "no triggers or subscriptions, cancelling watch.  "
-        "Set idle_reap_age_seconds in your .watchmanconfig to control "
-        "this behavior\n",
-        root->root_path->len, root->root_path->buf, root->idle_reap_age);
-    return true;
-  }
-
-  root->last_reap_timestamp = now;
-
-  return false;
-}
-
 void w_root_addref(w_root_t *root)
 {
   w_refcnt_add(&root->refcnt);
