+++ winbuild/Makefile
@@ -98,6 +98,7 @@ SRCS=\
 	root/lock.c       \
 	root/notifythread.c       \
 	root/poison.c       \
+	root/reap.c       \
 	root/resolve.c       \
 	root/stat.c       \
 	root/symlink.c       \
